requirements.txt
asgiref==3.6.0
catboost==1.1.1
contourpy==1.0.7
cycler==0.11.0
Django==4.2
django-cors-headers==3.10.0
djangorestframework==3.14.0
fonttools==4.39.3
graphviz==0.20.1
gunicorn==20.1.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
plotly==5.14.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.1.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
threadpoolctl==3.1.0
tzdata==2023.3
xgboost==1.7.4
